requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
WTForms
email_validator
python-dotenv
Flask-Migrate
pdf2image
pytesseract
langdetect
deep-translator
Werkzeug